chore: dto 한 클래스로 모으기
