+++ withaeng-domain/src/main/kotlin/com/withaeng/domain/accompany/AccompanyRepositoryCustom.kt
@@ -1,5 +1,6 @@
 package com.withaeng.domain.accompany
 
+import com.withaeng.domain.accompany.dto.FindAccompanyDto
 import com.withaeng.domain.accompany.dto.SearchAccompanyDto
 import com.withaeng.domain.accompany.dto.SearchAccompanyQuery
 import org.springframework.data.domain.Page
